requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.5
h11==0.14.0
httptools==0.6.4
idna==3.10
mysql-connector-python==9.1.0
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1

sqlmodel~=0.0.22
SQLAlchemy~=2.0.36
passlib~=1.7.4
PyJWT~=2.10.1
aioredis~=2.0.1